Tested add, delete, update, max, min, second max, second min, search user, top three successfully

+++ src/main/java/com/faisal/Model/Employee.java
@@ -16,7 +16,7 @@ public class Employee {
 	@NonNull
 	private String emp_name;
 	@NonNull
-	private float emp_salary ;
+	private double emp_salary ;
 	@NonNull
 	private String emp_country;
 	@NonNull
